+++ src/data7/__init__.py
@@ -1,1 +1,5 @@
 """Data7 root module."""
+
+from . import config
+
+__all__ = ["config"]