+++ src/data7/config.py
@@ -1,11 +1,15 @@
 """Data7 configuration module."""
 
+from typing import List
+
 from dynaconf import Dynaconf
+
+SETTINGS_FILES: List[str] = ["settings.yaml", ".secrets.yaml", "data7.yaml"]
 
 settings = Dynaconf(
     envvar_prefix="DATA7",
     root_path=".",
-    settings_files=["settings.yaml", ".secrets.yaml", "data7.yaml"],
+    settings_files=SETTINGS_FILES,
     environments=True,
     load_dotenv=True,
 )